TestTree refactoring
